Disease/requirements.txt
Flask
tensorflow
gunicorn
numpy
pillow
opencv-python
scikit-image